Updated components to more accurately be named
