+++ src/App.tsx
@@ -1,11 +1,11 @@
 import React from "react";
-import CreateUser from "./components/create-user";
+import Login from "./components/login";
 
 const App: React.FC = () => {
   // currently only have a form to create a user
   return (
     <>
-      <CreateUser />
+      <Login />
     </>
   );
 };
